LayoutTests/imported/w3c/web-platform-tests/streams/piping/general-expected.txt

CONSOLE MESSAGE: line 12: Unhandled Promise Rejection: TypeError: The WritableStream.ready getter can only be used on instances of WritableStream
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 2765: TypeError: undefined is not an object (evaluating 'e.error')
CONSOLE MESSAGE: Unhandled Promise Rejection: undefined
CONSOLE MESSAGE: line 2765: TypeError: undefined is not an object (evaluating 'e.error')
CONSOLE MESSAGE: Unhandled Promise Rejection: undefined
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 2765: TypeError: undefined is not an object (evaluating 'e.error')
CONSOLE MESSAGE: Unhandled Promise Rejection: undefined
CONSOLE MESSAGE: line 2765: TypeError: undefined is not an object (evaluating 'e.error')
CONSOLE MESSAGE: Unhandled Promise Rejection: undefined
CONSOLE MESSAGE: line 2765: TypeError: undefined is not an object (evaluating 'e.error')
CONSOLE MESSAGE: Unhandled Promise Rejection: undefined

Harness Error (FAIL), message = The WritableStream.ready getter can only be used on instances of WritableStream

FAIL Piping must lock both the ReadableStream and WritableStream assert_false: sanity check: the WritableStream must not start locked expected false got undefined
FAIL Piping finishing must unlock both the ReadableStream and WritableStream assert_false: the WritableStream must become unlocked expected false got undefined
FAIL pipeTo must check the brand of its ReadableStream this value Can only call ReadableStream.getReader on instances of ReadableStream
PASS pipeTo must check the brand of its WritableStream argument 
FAIL pipeTo must fail if the ReadableStream is locked, and not lock the WritableStream assert_false: sanity check: the WritableStream does not start locked expected false got undefined
FAIL pipeTo must fail if the WritableStream is locked, and not lock the ReadableStream ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)
FAIL Piping from a ReadableStream from which lots of chunks are synchronously readable promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Piping from a ReadableStream for which a chunk becomes asynchronously readable after the pipeTo assert_array_equals: close must always be called with the controller lengths differ, expected 1 got 0
PASS an undefined rejection from pull should cause pipeTo() to reject when preventAbort is true 
PASS an undefined rejection from pull should cause pipeTo() to reject when preventAbort is false 
PASS an undefined rejection from write should cause pipeTo() to reject when preventCancel is true 
PASS an undefined rejection from write should cause pipeTo() to reject when preventCancel is false 
